+++ src/wallet/rpcdump.cpp
@@ -894,10 +894,10 @@ UniValue dumpwallet(const Config &config, const JSONRPCRequest &request) {
     // add the base58check encoded extended master if the wallet uses HD
     CKeyID seed_id = pwallet->GetHDChain().seed_id;
     if (!seed_id.IsNull()) {
-        CKey key;
-        if (pwallet->GetKey(seed_id, key)) {
+        CKey seed;
+        if (pwallet->GetKey(seed_id, seed)) {
             CExtKey masterKey;
-            masterKey.SetSeed(key.begin(), key.size());
+            masterKey.SetSeed(seed.begin(), seed.size());
 
             file << "# extended private masterkey: " << EncodeExtKey(masterKey)
                  << "\n\n";
@@ -917,11 +917,11 @@ UniValue dumpwallet(const Config &config, const JSONRPCRequest &request) {
                                          strLabel)) {
                 file << strprintf("label=%s", strLabel);
             } else if (keyid == seed_id) {
-                file << "hdmaster=1";
+                file << "hdseed=1";
             } else if (mapKeyPool.count(keyid)) {
                 file << "reserve=1";
-            } else if (pwallet->mapKeyMetadata[keyid].hdKeypath == "m") {
-                file << "inactivehdmaster=1";
+            } else if (pwallet->mapKeyMetadata[keyid].hdKeypath == "s") {
+                file << "inactivehdseed=1";
             } else {
                 file << "change=1";
             }