+++ test/functional/wallet_keypool.py
@@ -19,8 +19,10 @@ class KeyPoolTest(BitcoinTestFramework):
         addr_before_encrypting_data = nodes[
             0].getaddressinfo(addr_before_encrypting)
         wallet_info_old = nodes[0].getwalletinfo()
+        assert_equal(wallet_info_old['hdseedid'],
+                     wallet_info_old['hdmasterkeyid'])
         assert(addr_before_encrypting_data[
-               'hdmasterkeyid'] == wallet_info_old['hdmasterkeyid'])
+               'hdseedid'] == wallet_info_old['hdseedid'])
 
         # Encrypt wallet and wait to terminate
         nodes[0].node_encrypt_wallet('test')
@@ -30,9 +32,10 @@ class KeyPoolTest(BitcoinTestFramework):
         addr = nodes[0].getnewaddress()
         addr_data = nodes[0].getaddressinfo(addr)
         wallet_info = nodes[0].getwalletinfo()
+        assert_equal(wallet_info['hdseedid'], wallet_info['hdmasterkeyid'])
         assert(addr_before_encrypting_data[
-               'hdmasterkeyid'] != wallet_info['hdmasterkeyid'])
-        assert(addr_data['hdmasterkeyid'] == wallet_info['hdmasterkeyid'])
+               'hdseedid'] != wallet_info['hdseedid'])
+        assert(addr_data['hdseedid'] == wallet_info['hdseedid'])
         assert_raises_rpc_error(
             -12, "Error: Keypool ran out, please call keypoolrefill first", nodes[0].getnewaddress)